refactor: 디렉토리 구조를 비즈니스 로직별로 변경
가독성 및 개발시 접근성을 위해 chatting, matching, member관련 비즈니스 도메인별로 각 디렉토리를 병햡
jwt와 구별을 위해 Authentication.java => EmailAuthentication.java 변경

resolves: #41
