--- src/main/java/com/aliens/friendship/authentication/repository/AuthenticationRepository.java
@@ -1,7 +1,0 @@
-package com.aliens.friendship.authentication.repository;
-
-import com.aliens.friendship.authentication.domain.Authentication;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AuthenticationRepository extends JpaRepository<Authentication, Integer> {
-}
