--- src/main/java/com/aliens/friendship/blockingInfo/repository/BlockingInfoRepository.java
@@ -1,7 +1,0 @@
-package com.aliens.friendship.blockingInfo.repository;
-
-import com.aliens.friendship.blockingInfo.domain.BlockingInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BlockingInfoRepository extends JpaRepository<BlockingInfo, Integer> {
-}
