+++ src/main/java/com/aliens/friendship/chatting/controller/dto/ChatMessage.java
@@ -1,0 +1,21 @@
+package com.aliens.friendship.chatting.controller.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class ChatMessage {
+
+    private Integer roomId;
+    private String sender;
+    private String message;
+    private LocalDateTime sendDate;
+
+}
