+++ src/main/java/com/aliens/friendship/chatting/controller/dto/RoomForm.java
@@ -1,0 +1,8 @@
+package com.aliens.friendship.chatting.controller.dto;
+
+import lombok.Data;
+
+@Data
+public class RoomForm {
+    private String name;
+}
