+++ src/main/java/com/aliens/friendship/chatting/repository/ChattingRepository.java
@@ -1,8 +1,8 @@
 package com.aliens.friendship.chatting.repository;
 
 import com.aliens.friendship.chatting.domain.Chatting;
-import com.aliens.friendship.chattingRoom.domain.ChattingRoom;
-import com.aliens.friendship.matchingParticipant.domain.MatchingParticipant;
+import com.aliens.friendship.chatting.domain.ChattingRoom;
+import com.aliens.friendship.matching.domain.MatchingParticipant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
