--- src/main/java/com/aliens/friendship/chattingRoom/controller/RoomForm.java
@@ -1,8 +1,0 @@
-package com.aliens.friendship.chattingRoom.controller;
-
-import lombok.Data;
-
-@Data
-public class RoomForm {
-    private String name;
-}
