--- src/main/java/com/aliens/friendship/chattingRoom/repository/ChattingRoomRepository.java
@@ -1,7 +1,0 @@
-package com.aliens.friendship.chattingRoom.repository;
-
-import com.aliens.friendship.chattingRoom.domain.ChattingRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChattingRoomRepository extends JpaRepository<ChattingRoom, Integer> {
-}
