+++ src/main/java/com/aliens/friendship/emailAuthentication/repository/EmailAuthenticationRepository.java
@@ -1,0 +1,7 @@
+package com.aliens.friendship.emailAuthentication.repository;
+
+import com.aliens.friendship.emailAuthentication.domain.EmailAuthentication;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EmailAuthenticationRepository extends JpaRepository<EmailAuthentication, Integer> {
+}
