+++ src/main/java/com/aliens/friendship/matching/repository/BlockingInfoRepository.java
@@ -1,0 +1,7 @@
+package com.aliens.friendship.matching.repository;
+
+import com.aliens.friendship.matching.domain.BlockingInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BlockingInfoRepository extends JpaRepository<BlockingInfo, Integer> {
+}
