+++ src/main/java/com/aliens/friendship/matching/repository/LanguageRepository.java
@@ -1,0 +1,7 @@
+package com.aliens.friendship.matching.repository;
+
+import com.aliens.friendship.matching.domain.Language;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LanguageRepository extends JpaRepository<Language, Integer> {
+}
