--- src/main/java/com/aliens/friendship/matchingParticipant/MatchingParticipantInfo.java
@@ -1,17 +1,0 @@
-package com.aliens.friendship.matchingParticipant;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MatchingParticipantInfo {
-
-    private int memberId;
-    private int answer; // 1 또는 2 값
-    private int language;
-}
