--- src/main/java/com/aliens/friendship/matchingParticipant/repository/MatchingParticipantRepository.java
@@ -1,7 +1,0 @@
-package com.aliens.friendship.matchingParticipant.repository;
-
-import com.aliens.friendship.matchingParticipant.domain.MatchingParticipant;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MatchingParticipantRepository extends JpaRepository<MatchingParticipant, Integer> {
-}
