+++ src/main/java/com/aliens/friendship/member/repository/NationalityRepository.java
@@ -1,0 +1,7 @@
+package com.aliens.friendship.member.repository;
+
+import com.aliens.friendship.member.domain.Nationality;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NationalityRepository extends JpaRepository<Nationality, Integer> {
+}
