--- src/main/java/com/aliens/friendship/question/repository/QuestionRepository.java
@@ -1,8 +1,0 @@
-package com.aliens.friendship.question.repository;
-
-import com.aliens.friendship.question.domain.Question;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface QuestionRepository extends JpaRepository<Question, Integer> {
-    Question findByIsSelected(byte isSelected);
-}
